+++ src/avz_click.cpp
@@ -42,8 +42,7 @@ void AShovel(int row, float col, int targetType)
         return;
     }
 
-    auto&& pattern = __aig.loggerPtr->GetPattern();
-    __aig.loggerPtr->Info("Shovel (" + pattern + ", " + pattern + ")", row, col);
+    aLogger->Info("Shovel {}", APosition(row, col));
     for (int i = 0; i < 10; ++i) {
         ALeftClick(x, y);
     }
@@ -67,13 +66,10 @@ void AShovel(const std::vector<AShovelPosition>& lst)
 void AGridToCoordinate(int row, float col, int& x, int& y)
 {
     int tCol = std::clamp(int(col + 0.5), 1, 10);
-    x = col * 80;
+    x = int(col * 80.0 + 1e-3);
     y = AAsm::GridToOrdinate(row - 1, tCol - 1) + 40;
-    if (col <= 0 || row <= 0 || x < 0 || x >= 800 || y < 0 || y >= 600) {
-        AGetInternalLogger()->Error("您输入的格子位置参数 : ("
-                + std::to_string(row) + ", " + AGetInternalLogger()->GetPattern()
-                + ") 已溢出, 已帮您自动调整为边界值",
-            col);
+    if (row <= 0 || x < 0 || x >= 800 || y < 0 || y >= 600) {
+        aLogger->Error("您输入的格子位置参数: {} 已溢出, 已帮您自动调整为边界值", APosition(row, col));
     }
     x = std::clamp(x, 0, 800 - 1);
     y = std::clamp(y, 0, 600 - 1);